Use .jsx for JSX files, stop using webpack magic
- 'Urls' is just a global variable injected through Django-templates,
  there is a js-library we could use, but that would have be bad for
  performance since you have to request a .json file at runtime, and
  probably use async, which is a lot of code-changes.
