+++ assets/careeropportunity/components/InfoOnOpportunity.jsx
