+++ assets/careeropportunity/containers/DetailView.jsx
