+++ assets/core/LoadingComponent/index.jsx
