+++ assets/dashboard/common/DashboardPanel.jsx
