+++ assets/dashboard/hobbies/components/HobbyDetails.jsx
