+++ assets/dashboard/hobbies/index.jsx
