+++ assets/dashboard/resources/Router.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter as ReactRouter, Switch, Route } from 'react-router-dom';
-import Urls from 'urls';
+import Urls from 'common/utils/django_reverse';
 
 import ResourcesList from './containers/ResourcesList';
 import NewResource from './containers/NewResource';
