+++ assets/dashboard/resources/components/ResourceDetails.jsx
