+++ assets/dashboard/resources/containers/NewResource.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import Cookies from 'js-cookie';
-import Urls from 'urls';
+import Urls from 'common/utils/django_reverse';
 
 import ResourceForm from '../components/ResourceForm';
 import ResourceDetails from '../components/ResourceDetails';
