+++ assets/dashboard/resources/index.jsx
