+++ assets/frontpage/articles/components/MainArticle.jsx
