+++ assets/frontpage/articles/components/SmallArticle.jsx
