+++ assets/frontpage/events/components/Event.jsx
