+++ assets/frontpage/events/components/EventImage.jsx
