+++ assets/frontpage/events/components/Events.jsx
