+++ assets/frontpage/events/components/SmallEvent.jsx
